+++ app/client/App.vue
@@ -76,6 +76,7 @@ export default {
   --color-red: #ed7a66;
   --color-grey: #757575;
   --color-green: #98ff7d;
+  --color-yellow: #efeba2;
 
   --font-main: open-sans, sans-serif;
   --font-headings: grandma, serif;
@@ -208,10 +209,6 @@ strong {
   font-weight: bolder;
 }
 
-main {
-  margin: var(--spacing-double);
-}
-
 p {
   margin-bottom: var(--spacing);
 }
@@ -222,7 +219,7 @@ p {
 }
 
 .container > main {
-  flex: 2;
+  flex: 3;
 }
 
 .box-container {