+++ app/client/modules/ActiveUserList/index.vue
@@ -1,0 +1,3 @@
+<template
+  ><div>Active User List Goes Here...</div></template
+>