+++ app/client/views/Home.vue
@@ -1,16 +1,19 @@
 <template>
-  <main><NotebookList :notebooks="allNotebooks" /></main>
+  <main>
+    <MainNav />
+    <ActiveUserList />
+    <StoryFeed />
+  </main>
 </template>
 
 <script>
-import { mapActions, mapGetters } from 'vuex';
+import { mapActions } from 'vuex';
 
 export default {
   components: {
-    NotebookList: () => import('@client/modules/NotebookList')
-  },
-  computed: {
-    ...mapGetters(['allNotebooks'])
+    MainNav: () => import('@client/modules/MainNav'),
+    StoryFeed: () => import('@client/modules/StoryFeed'),
+    ActiveUserList: () => import('@client/modules/ActiveUserList')
   },
   watch: {
     $route() {